move balance functions to hook
