+++ src/library/Hooks/useActiveBalances/defaults.ts
@@ -1,0 +1,12 @@
+// Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+/* eslint-disable @typescript-eslint/no-unused-vars */
+
+import BigNumber from 'bignumber.js';
+import type { Balance } from 'contexts/Balances/types';
+
+export const defaultActiveBalance: Balance = {
+  free: new BigNumber(0),
+  reserved: new BigNumber(0),
+  frozen: new BigNumber(0),
+};
