+++ src/modals/Accounts/index.tsx
@@ -32,7 +32,6 @@ import type {
 } from './types';
 import type { ImportedAccount } from '@polkadot-cloud/react/types';
 import { useActiveBalances } from 'library/Hooks/useActiveBalances';
-import type { MaybeAddress } from 'types';
 
 export const Accounts = () => {
   const { t } = useTranslation('modals');
@@ -54,26 +53,14 @@ export const Accounts = () => {
     useState<ImportedAccount[]>(accounts);
 
   // Listen to balance updates for entire accounts list.
-  const activeBalances = useActiveBalances({
+  const { getBalanceLocks } = useActiveBalances({
     accounts: localAccounts.map(({ address }) => address),
   });
-
-  // Gets an account's locks from `activeBalances`.
-  // TODO: this function can be moved inside the hook.
-  const getLocks = (address: MaybeAddress) => {
-    if (address) {
-      const maybeLocks = activeBalances[address]?.balances.locks;
-      if (maybeLocks) {
-        return maybeLocks;
-      }
-    }
-    return [];
-  };
 
   const stashes: string[] = [];
   // accumulate imported stash accounts
   for (const { address } of localAccounts) {
-    const locks = getLocks(address);
+    const locks = getBalanceLocks(address);
 
     // account is a stash if they have an active `staking` lock
     if (locks.find(({ id }) => id === 'staking')) {
